+++ gatsby-config.js
@@ -90,16 +90,6 @@ module.exports = {
 				},
 			},
 		},
-		{
-			resolve: "gatsby-source-gravityforms",
-			options: {
-				baseUrl: `https://${process.env.WPC_WORDPRESS}`,
-				api: {
-					key: process.env.WPC_GF_API_KEY,
-					secret: process.env.WPC_GF_API_SECRET,
-				},
-			},
-		},
 		// this (optional) plugin enables Progressive Web App + Offline functionality
 		// To learn more, visit: https://gatsby.dev/offline
 		// `gatsby-plugin-offline`,
