+++ package.json
@@ -25,9 +25,9 @@
     "gatsby-plugin-sharp": "^2.4.0",
     "gatsby-plugin-sitemap": "^2.3.5",
     "gatsby-source-filesystem": "^2.1.35",
-    "gatsby-source-gravityforms": "^1.0.15",
     "gatsby-source-wordpress": "^3.1.47",
     "gatsby-transformer-sharp": "^2.3.1",
+    "iframe-resizer": "^4.2.10",
     "node-fetch": "^2.6.0",
     "postcss-nested": "^4.2.1",
     "prop-types": "^15.7.2",
